% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_annual}
\alias{aggregate_annual}
\title{Annual aggregation}
\usage{
aggregate_annual(x, ...)
}
\arguments{
\item{x}{catchment}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Convenience functions
}
