% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_daily}
\alias{aggregate_daily}
\title{Daily aggregation}
\usage{
aggregate_daily(x, ...)
}
\arguments{
\item{x}{tsibble.}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Convenience function to aggregate catchment data to daily temporal
resolution.
}
