% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_hourly}
\alias{aggregate_hourly}
\title{Hourly aggregation}
\usage{
aggregate_hourly(x, ...)
}
\arguments{
\item{x}{tsibble.}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Convenience function to aggregate catchment data to hourly temporal
resolution.
}
