% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_monthly}
\alias{aggregate_monthly}
\title{Monthly aggregation}
\usage{
aggregate_monthly(x, ...)
}
\arguments{
\item{x}{catchment}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Convenience functions
}
