% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregation_cs}
\alias{aggregation_cs}
\title{Aggregate rainfall data}
\usage{
aggregation_cs(
  x,
  timescale = set_units(60, "s"),
  bucket = set_units(0.2, "mm"),
  mintip = TRUE,
  halves = TRUE,
  ...
)
}
\arguments{
\item{x}{rain_gauge.}

\item{timescale}{units. Aggregation interval in seconds \link{original - minutes}}

\item{bucket}{units. Rain gauge bucket volume in mm.}

\item{mintip}{logical. Whether to aggregate at 1-min intervals
prior to interpolation.}

\item{halves}{logical. Whether to add zero rates at estimated
event endpoints.}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Aggregate rainfall data with cubic spline interpolation.
}
