% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{baseflow}
\alias{baseflow}
\title{Compute baseflow}
\usage{
baseflow(Date, Q, method = "UKIH", ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{x}{catchment.}
}
\value{

}
\description{
Implementation of UKIH baseflow separation algorithm
}
\examples{
\dontrun{
sum(1:10)
}
}
\references{
Gustard, A.; Bullock, A.; Dixon, J. M.. 1992 Low flow estimation in
the United Kingdom. Wallingford, Institute of Hydrology, 88pp. (IH
Report No.108). http://nora.nerc.ac.uk/id/eprint/6050 (Accessed
June 2022)

Chapman, T. (1999), A comparison of algorithms for stream flow
recession and baseflow separation. Hydrol. Process., 13: 701-714.
https://doi.org/10.1002/(SICI)1099-1085(19990415)13:5<701::AID-HYP774>3.0.CO;2-2
(Accessed June 2022)
}
