% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{baseflow_uk}
\alias{baseflow_uk}
\title{Compute baseflow}
\usage{
baseflow_uk(Date, Q, ...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
\dontrun{
sum(1:10)
}
}
