% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{compute_baseflow}
\alias{compute_baseflow}
\title{Compute baseflow}
\usage{
compute_baseflow(x, method = c("UKIH", "Chapman1999"), ...)
}
\arguments{
\item{x}{catchment.}

\item{method}{Character. One or both of 'UKIH' and 'Chapman1999'.}

\item{...}{Additional arguments.}
}
\value{

}
\description{
Implementation of two baseflow separation algorithms.
}
\references{
Gustard, A.; Bullock, A.; Dixon, J. M.. 1992 Low flow estimation in
the United Kingdom. Wallingford, Institute of Hydrology, 88pp. (IH
Report No.108). http://nora.nerc.ac.uk/id/eprint/6050 (Accessed
June 2022)

Chapman, T. (1999), A comparison of algorithms for stream flow
recession and baseflow separation. Hydrol. Process., 13: 701-714.
https://doi.org/10.1002/(SICI)1099-1085(19990415)13:5<701::AID-HYP774>3.0.CO;2-2
(Accessed June 2022)
}
