% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{compute_indices}
\alias{compute_indices}
\title{Compute indices}
\usage{
compute_indices(x, area, ...)
}
\arguments{
\item{x}{catchment.}

\item{area}{units.}

\item{...}{Additional arguments.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
\dontrun{
sum(1:10)
}
}
