% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depure.R
\name{depure}
\alias{depure}
\title{Remove consecutive tips}
\usage{
depure(x, ...)
}
\arguments{
\item{x}{tipping_bucket_rain_gauge}

\item{...}{Additional arguments.}
}
\value{
tibble
}
\description{
Remove consecutive tips above maximum intensity
}
