% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{discharge}
\alias{discharge}
\title{Create discharge object}
\usage{
discharge(x)
}
\arguments{
\item{x}{Data frame.}
}
\value{
A discharge object.
}
\description{
Create iMHEA discharge object.
}
\examples{
\dontrun{
sum(1:10)
}
}
