% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Aggregate rainfall data}
\usage{
fill_gaps(Date1, P1, Date2, P2, cutend = FALSE, ...)
}
\arguments{
\item{Date1}{rain_gauge.}

\item{P1}{rain_gauge.}

\item{Date2}{rain_gauge.}

\item{P2}{rain_gauge.}

\item{cutend}{logical.}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Aggregate rainfall data with cubic spline interpolation.
}
