% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voids.R
\name{identify_voids}
\alias{identify_voids}
\title{Void intervals}
\usage{
identify_voids(x, ...)
}
\arguments{
\item{x}{tsibble}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Determine void intervals.
}
