% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{rain_gauge}
\alias{rain_gauge}
\title{Load rain_gauge object}
\usage{
rain_gauge(
  file,
  tz = "Etc/GMT-5",
  lat = NA,
  lon = NA,
  gauge_type,
  bucket_volume = 0.2,
  date_column = "Date",
  event_column = "Event mm"
)
}
\arguments{
\item{file}{Character. Path to input precipitation data.}

\item{tz}{Character. Timezone of recorded times. Defaults to
"Etc/GMT-5" (i.e. timezone of Peru and Ecuador)}

\item{lat}{Numeric.}

\item{lon}{Numeric.}

\item{gauge_type}{Character.}

\item{bucket_volume}{Numeric vector.}

\item{date_column}{Character. The name of the date column. Default
is "Date"}

\item{event_column}{Character. The name of the event column.
Default is "Event mm"}

\item{...}{Additional arguments.}
}
\value{
A precipitation object.
}
\description{
Load iMHEA precipitation object from file.
}
\examples{
\dontrun{
sum(1:10)
}
}
