% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_gauge.R
\name{stream_gauge}
\alias{stream_gauge}
\alias{is_stream_gauge}
\title{Load stream_gauge object}
\usage{
stream_gauge(
  x,
  id,
  date_column = "Date",
  date_format = "\%d/\%m/\%Y \%H:\%M:\%S",
  tz = "Etc/GMT-5",
  discharge_column = "Flow l/s",
  discharge_units,
  level_column = NA,
  level_units,
  flag_column = "Flag",
  raw = TRUE,
  ...
)

is_stream_gauge(x)
}
\arguments{
\item{x}{tibble. Raw data.}

\item{id}{Character. Gauge ID.}

\item{date_column}{Character. Name of column holding time information.}

\item{date_format}{Character. Date format.}

\item{tz}{Character. Timezone of recorded times. Defaults to
"Etc/GMT-5" (i.e. timezone of Peru and Ecuador)}

\item{flag_column}{Character. The name of the flag column. Default
is "Flag"}

\item{...}{Additional arguments.}

\item{discharge_column.}{Character. Name of column holding discharge data.
Default is "Flow l/s"}

\item{discharge_units.}{Character. Discharge units (e.g. l/s, m3/s).}

\item{level_column.}{Character. Name of column holding level data.}

\item{level_units.}{Character. Level units (e.g. m, mm).}
}
\value{
A stream_gauge object.
}
\description{
Load stream_gauge object
}
\examples{

q_raw <- read_csv(
  system.file("extdata", "LLO/iMHEA_LLO_01_HI_01_raw.csv", package = "imhea"),
  show_col_types = FALSE
)

q <-
  q_raw \%>\%
  stream_gauge(
    id = "LLO_01_HI_01",
    discharge_units = "l/s",
    level_column = "Level cm",
    level_units = "cm"
  )

is_stream_gauge(q)

}
