% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_gauge.R
\name{tipping_bucket_rain_gauge}
\alias{tipping_bucket_rain_gauge}
\alias{is_rain_gauge}
\title{Load rain_gauge object}
\usage{
tipping_bucket_rain_gauge(
  x,
  id,
  date_column = "Date",
  date_format = "\%d/\%m/\%Y \%H:\%M:\%S",
  tz = "Etc/GMT-5",
  event_column = "Event mm",
  event_units,
  flag_column = "Flag",
  ...
)

is_rain_gauge(x)
}
\arguments{
\item{x}{tibble. Raw data.}

\item{id}{Character. Gauge ID.}

\item{date_column}{Character. Name of column holding time information.}

\item{date_format}{Character. Date format.}

\item{tz}{Character. Timezone of recorded times. Defaults to
"Etc/GMT-5" (i.e. timezone of Peru and Ecuador)}

\item{event_column}{Character. The name of the event column.
Default is "Event mm"}

\item{event_units}{Character. Precipitation units (e.g. mm).}

\item{flag_column}{Character. The name of the flag column. Default
is "Flag"}

\item{...}{Additional arguments.}
}
\value{
A rain_gauge object.
}
\description{
Load rain_gauge object
}
\examples{

p_raw = read_csv(
  system.file("extdata", "LLO/iMHEA_LLO_01_PO_01_raw.csv", package = "imhea"),
  show_col_types = FALSE
)

p <- p_raw \%>\% tipping_bucket_rain_gauge(id = "LLO_01_P0_01", event_units = "mm")

is_rain_gauge(p)

}
