% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{update_indices}
\alias{update_indices}
\title{Compute indices}
\usage{
update_indices(x, ...)
}
\arguments{
\item{x}{catchment.}

\item{...}{Additional arguments.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
\dontrun{
sum(1:10)
}
}
