% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate data}
\usage{
aggregate(x, timescale, ...)
}
\arguments{
\item{x}{tsibble.}

\item{timescale}{units.}

\item{...}{Additional arguments.}
}
\value{
tsibble
}
\description{
Aggregate rainfall data with cubic spline interpolation.
}
