% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.R
\name{catchment}
\alias{catchment}
\title{Create catchment object}
\usage{
catchment(area, discharge, precipitation)
}
\arguments{
\item{area}{Numeric. Catchment area \link{km2}}

\item{precipitation}{}
}
\value{
A catchment object
}
\description{
Create iMHEA catchment object.
}
\examples{
\dontrun{
sum(1:10)
}
}
