% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.R
\name{catchment}
\alias{catchment}
\title{Create catchment object}
\usage{
catchment(q, ..., id, area)
}
\arguments{
\item{q}{stream_gauge}

\item{...}{rain_gauge objects.}

\item{id}{character.}

\item{area}{units}
}
\value{
A catchment object
}
\description{
Create iMHEA catchment object.
}
\examples{
\dontrun{
sum(1:10)
}
}
