% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Compute indices}
\usage{
indices(Date, P, Q, A, ...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
\dontrun{
sum(1:10)
}
}
